Disabled embree due to race conditions Added CudaHostBuffer class to wrap page-locked host (CPU) memory Updated NewCudaTwoLevelRayTracer to use the new wrapped CUDA objects Updated CudaImage class to remove alignment options Updated todo.txt Fixed warnings in Distribution1 and TrialData
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@2865 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ utils/Distribution1.h
@@ -25,7 +25,7 @@ public:
 
 	Distribution1(const std::vector<Real>& pdf, float guide_table_size_scale = 1)
 	{
-		const uint32 table_size = (uint32)pdf.size();
+		const size_t table_size = pdf.size();
 
 		cdfTable.resize(table_size);
 		guideTable.resize(table_size * guide_table_size_scale);
@@ -44,7 +44,7 @@ public:
 		}
 
 		// Build the guide table.
-		for(uint32 i = 0, j = 0; i < guideTable.size(); ++i)
+		for(size_t i = 0, j = 0; i < guideTable.size(); ++i)
 		{
 			// Find the largest j st. CDF[j] < target CDF[i]
 			const double target_cdf = i / (double)guideTable.size();
